Refractored System class to User,Device
