--- backend/src/main/java/slabs/tech/avti/persistance/model/Stall.java
@@ -1,68 +1,0 @@
-package slabs.tech.avti.persistance.model;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.Table;
-
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@Table(name = "stall")
-@EntityListeners(AuditingEntityListener.class)
-public class Stall {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "stall_generator")
-	@SequenceGenerator(name = "stall_generator", sequenceName = "stall_id_seq", allocationSize = 1)
-	@Column(name = "id", updatable = false, nullable = false)
-	private long id;
-
-	@Column(name = "user_name", nullable = false, unique = true)
-	private String user_name;
-
-	@Column(name = "latitude", nullable = false, unique = true)
-	private double latitude;
-
-	@Column(name = "longitude", nullable = false, unique = true)
-	private double longitude;
-
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
-	public String getUser_name() {
-		return user_name;
-	}
-
-	public void setUser_name(String user_name) {
-		this.user_name = user_name;
-	}
-
-	public double getLatitude() {
-		return latitude;
-
-	}
-
-	public double getLongitude() {
-		return longitude;
-	}
-
-	public void setLongitude(double longitude) {
-		this.longitude = longitude;
-	}
-
-	public void setLatitude(double latitude) {
-		this.latitude = latitude;
-	}
-
-	
-}
