+++ backend/src/main/java/slabs/tech/avti/persistance/repo/DeviceRepository.java
@@ -1,0 +1,9 @@
+package slabs.tech.avti.persistance.repo;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import slabs.tech.avti.persistance.model.Device;
+
+public interface DeviceRepository extends JpaRepository<Device, Long> {
+	  
+}
