--- backend/src/main/java/slabs/tech/avti/persistance/repo/DriverRepository.java
@@ -1,9 +1,0 @@
-package slabs.tech.avti.persistance.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import slabs.tech.avti.persistance.model.Driver;
-
-public interface DriverRepository extends JpaRepository<Driver, Long> {
-	  
-}
