--- backend/src/main/java/slabs/tech/avti/persistance/repo/StallRepository.java
@@ -1,9 +1,0 @@
-package slabs.tech.avti.persistance.repo;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import slabs.tech.avti.persistance.model.Stall;
-
-public interface StallRepository extends JpaRepository<Stall, Long> {
-	  
-}
