+++ backend/src/main/java/slabs/tech/avti/web/DeviceController.java
@@ -13,15 +13,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import slabs.tech.avti.persistance.model.Vehicle;
-import slabs.tech.avti.persistance.repo.VehicleRepository;
+import slabs.tech.avti.persistance.model.Device;
+import slabs.tech.avti.persistance.repo.DeviceRepository;
 
 @RestController
 @RequestMapping("/vehicles")
-public class VehicleController {
+public class DeviceController {
 
 	@Autowired
-	private VehicleRepository vehicleRepository;
+	private DeviceRepository vehicleRepository;
 
 	/**
 	 * Get all Vehicles list.
@@ -29,7 +29,7 @@ public class VehicleController {
 	 * @return the list
 	 */
 	@GetMapping // GET Method for reading operation
-	public List<Vehicle> getAllVehicles() {
+	public List<Device> getAllVehicles() {
 		return vehicleRepository.findAll();
 	}
 
@@ -41,9 +41,9 @@ public class VehicleController {
 	 * @throws Exception
 	 */
 	@GetMapping("/{id}") // GET Method for Read operation
-	public ResponseEntity<Vehicle> getVehicleById(@PathVariable(value = "id") Long vehicleId) throws Exception {
+	public ResponseEntity<Device> getVehicleById(@PathVariable(value = "id") Long vehicleId) throws Exception {
 
-		Vehicle vehicle = vehicleRepository.findById(vehicleId)
+		Device vehicle = vehicleRepository.findById(vehicleId)
 				.orElseThrow(() -> new Exception("Vehicle " + vehicleId + " not found"));
 		return ResponseEntity.ok().body(vehicle);
 	}
@@ -55,7 +55,7 @@ public class VehicleController {
 	 * @return the vehicle
 	 */
 	@PostMapping // POST Method for Create operation
-	public Vehicle createVehicle(@Valid @RequestBody Vehicle vehicle) {
+	public Device createVehicle(@Valid @RequestBody Device vehicle) {
 		return vehicleRepository.save(vehicle);
 	}
